+++ src/main/java/absolutelyaya/ultracraft/util/InventoryUtil.java
@@ -1,8 +1,13 @@
 package absolutelyaya.ultracraft.util;
 
+import com.google.common.primitives.ImmutableIntArray;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.collection.DefaultedList;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 
 public class InventoryUtil
 {
@@ -19,5 +24,31 @@ public class InventoryUtil
 			}
 		}
 		return false;
+	}
+	
+	public static boolean containsItem(DefaultedList<ItemStack> inv, Item item)
+	{
+		return containsItem(inv, item, 1);
+	}
+	
+	public static int getFirstSlotWithItem(DefaultedList<ItemStack> inv, Item item)
+	{
+		for (int i = 0; i < inv.size(); i++)
+		{
+			if(inv.get(i).isOf(item))
+				return i;
+		}
+		return -1;
+	}
+	
+	public static int[] getAllSlowsWithItem(DefaultedList<ItemStack> inv, Item item)
+	{
+		ImmutableIntArray.Builder builder = ImmutableIntArray.builder();
+		for (int i = 0; i < inv.size(); i++)
+		{
+			if(inv.get(i).isOf(item))
+				builder.add(i);
+		}
+		return builder.build().stream().toArray();
 	}
 }